backend/requirements.txt
Django>=3.2.3,<3.3
psycopg2>=2.8.6,<2.9
uWSGI>=2.0.19.1,<2.1
djangorestframework
django-cors-headers
django-rest-swagger
drf_yasg
django-extensions
pandas
python-dateutil
chromadb